Numpad (numbers on a line) added to qwerty setup
When NumLock AND CapsLock are ON, the buttons at the touch-method's default positions, are turned into a line of numbers. 
The number line is in essence moved down two rows in the qwerty keyboard. 
This way numbers are more easily accessible, and both comma and period (, and . ) are accessible, as opposed to standard numpad keyboards. The operators are also more accessible , like the minus sign, division sign and multiplication sign. The pluss sign is finally located on the pointing finger, which is far more ergonomical and fast than moving the little finger on standard numpad keyboards.